move local source checking to under resolve_deployment_source
